requirements.txt
numpy
pandas
Pillow >= 10.1.10
scikit-learn
matplotlib